explict list of license strings
Also allows license to be an arrayref to indicate multiple
license may apply to the distribution.

All license strings are versioned where appropriate to do so.
This ensures we no longer have confusion over which version
license strings like 'gpl' or 'apache' refer to.  (This is a
break from version 1.X of the spec.)