Add OnlyCorePrereqs plugin to dist.ini for test phase
We depend on other modules at runtime that might not be shipped
with Perl, yet.  When OnlyCorePrereqs is fixed to allow "any version
of module Foo, as long as Foo is in core" we can change this
to check runtime as well.

Meanwhile, checking the test phase should be sufficient to guard
against non-runtime dependencies creeping in.