Improved fix for scalar meta-spec field
If 'meta-spec' is not a hash reference, this now treats it like a 1.0
spec (when 'meta-spec' was not defined).  It will convert 'meta-spec' to
a valid hash reference during conversion.  This allows possibly
'rescuing' a bad META file.

It also adds t/README-data.txt to clarify where to put data files (since
I'd forgotten myself and had to figure out what we meant with each
directory).